Rename Tab to GenericTabs
